test(validation): move hasError return false to throws validation handler test class
